+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - [NDMF] Logo to the NDMF Console.
+- [NDMF] Feature to set as a fallback avatar in the VQT Avatar Builder.
 - Add warning to suggest converting Unity constraints to VRChat constraints in `VQT Avatar Converter Settings`.
 - Build and publish for iOS platform (as well as Android).